third_party/WebKit/LayoutTests/ambient-light/multiple-event-listeners-expected.txt

Tests using multiple event handlers for the Device Light API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS deviceLightEvent.value is expectedEvent.value
PASS deviceLightEvent.value is expectedEvent.value
PASS deviceLightEvent.value is expectedEvent.value
PASS deviceLightEvent.value is expectedEvent.value
PASS successfullyParsed is true

TEST COMPLETE
